+++ src/uscxml/messages/Event.cpp
@@ -18,7 +18,10 @@
  */
 
 #include "uscxml/messages/Event.h"
-#include "uscxml/DOMUtils.h"
+#include "uscxml/dom/DOMUtils.h"
+#include <DOM/SAX2DOM/SAX2DOM.hpp>
+#include <SAX/helpers/DefaultHandler.hpp>
+#include <SAX/helpers/CatchErrorHandler.hpp>
 
 namespace uscxml {
 
@@ -97,12 +100,12 @@ Event Event::fromXML(const std::string& xmlString) {
 				Arabica::DOM::Element<std::string> payloadElem = (Arabica::DOM::Element<std::string>)payload;
 				Arabica::DOM::NodeList<std::string> properties = payloadElem.getElementsByTagName("scxml:property");
 				if (properties.getLength() > 0) {
-					for (int i = 0; i < properties.getLength(); i++) {
+					for (size_t i = 0; i < properties.getLength(); i++) {
 						if (HAS_ATTR_CAST(properties.item(i), "name")) {
 							std::string key = ATTR_CAST(properties.item(i), "name");
 							std::string value;
 							Arabica::DOM::NodeList<std::string> childs = properties.item(i).getChildNodes();
-							for (int j = 0; j < childs.getLength(); j++) {
+							for (size_t j = 0; j < childs.getLength(); j++) {
 								if (childs.item(j).getNodeType() == Arabica::DOM::Node_base::TEXT_NODE) {
 									value = childs.item(j).getNodeValue();
 									break;
@@ -120,7 +123,7 @@ Event Event::fromXML(const std::string& xmlString) {
 
 std::ostream& operator<< (std::ostream& os, const Event& event) {
 	std::string indent;
-	for (int i = 0; i < _dataIndentation; i++) {
+	for (size_t i = 0; i < _dataIndentation; i++) {
 		indent += "  ";
 	}
