+++ test/src/test-lifecycle.cpp
@@ -28,7 +28,7 @@ bool testIssue56();
 #ifdef HAS_EXECINFO_H
 void printBacktrace(void** array, int size) {
 	char** messages = backtrace_symbols(array, size);
-	for (int i = 0; i < size && messages != NULL; ++i) {
+	for (size_t i = 0; i < size && messages != NULL; ++i) {
 		std::cerr << "\t" << messages[i] << std::endl;
 	}
 	std::cerr << std::endl;
@@ -275,9 +275,9 @@ int main(int argc, char** argv) {
 			Interpreter interpreter = Interpreter::fromXML(xml, "");
 			interpreter.addMonitor(mon);
 
-            callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
-            callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
-            callBackSeq.push_back(USCXML_BEFOREENTERINGSTATE);
+			callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_BEFOREENTERINGSTATE);
 			callBackSeq.push_back(USCXML_AFTERENTERINGSTATE);
 
 			callBackSeq.push_back(USCXML_BEFOREMICROSTEP);
@@ -301,10 +301,10 @@ int main(int argc, char** argv) {
 			callBackSeq.push_back(USCXML_BEFORECOMPLETION);
 			callBackSeq.push_back(USCXML_AFTERCOMPLETION);
 
-            assert(interpreter.getState() == USCXML_INSTANTIATED);
-            assert(interpreter.step() == USCXML_INITIALIZED);
+			assert(interpreter.getState() == USCXML_INSTANTIATED);
+			assert(interpreter.step() == USCXML_INITIALIZED);
 			assert(interpreter.step() == USCXML_MICROSTEPPED);
-            assert(interpreter.step() == USCXML_MICROSTEPPED);
+			assert(interpreter.step() == USCXML_MICROSTEPPED);
 			assert(interpreter.step() == USCXML_FINISHED);
 			assert(callBackSeq.empty());
 		}
@@ -322,8 +322,8 @@ int main(int argc, char** argv) {
 			Interpreter interpreter = Interpreter::fromXML(xml, "");
 			interpreter.addMonitor(mon);
 
-            callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
-            callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
 			callBackSeq.push_back(USCXML_BEFOREENTERINGSTATE);
 			callBackSeq.push_back(USCXML_AFTERENTERINGSTATE);
 
@@ -340,13 +340,13 @@ int main(int argc, char** argv) {
 			callBackSeq.push_back(USCXML_AFTERCOMPLETION);
 
 			assert(interpreter.getState() == USCXML_INSTANTIATED);
-            assert(interpreter.step() == USCXML_INITIALIZED);
-            assert(interpreter.step() == USCXML_MICROSTEPPED);
+			assert(interpreter.step() == USCXML_INITIALIZED);
+			assert(interpreter.step() == USCXML_MICROSTEPPED);
 			assert(interpreter.step() == USCXML_FINISHED);
 			interpreter.reset();
 
-            callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
-            callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
 			callBackSeq.push_back(USCXML_BEFOREENTERINGSTATE);
 			callBackSeq.push_back(USCXML_AFTERENTERINGSTATE);
 
@@ -363,8 +363,8 @@ int main(int argc, char** argv) {
 			callBackSeq.push_back(USCXML_AFTERCOMPLETION);
 
 			assert(interpreter.getState() == USCXML_INSTANTIATED);
-            assert(interpreter.step() == USCXML_INITIALIZED);
-            assert(interpreter.step() == USCXML_MICROSTEPPED);
+			assert(interpreter.step() == USCXML_INITIALIZED);
+			assert(interpreter.step() == USCXML_MICROSTEPPED);
 			assert(interpreter.step() == USCXML_FINISHED);
 		}
 
@@ -387,8 +387,8 @@ int main(int argc, char** argv) {
 			Interpreter interpreter = Interpreter::fromXML(xml, "");
 			interpreter.addMonitor(mon);
 
-            callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
-            callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_BEFORETAKINGTRANSITION);
+			callBackSeq.push_back(USCXML_AFTERTAKINGTRANSITION);
 			callBackSeq.push_back(USCXML_BEFOREENTERINGSTATE);
 			callBackSeq.push_back(USCXML_BEFOREEXECUTINGCONTENT);
 			callBackSeq.push_back(USCXML_AFTEREXECUTINGCONTENT);
@@ -417,11 +417,11 @@ int main(int argc, char** argv) {
 			callBackSeq.push_back(USCXML_BEFORECOMPLETION);
 			callBackSeq.push_back(USCXML_AFTERCOMPLETION);
 
-            assert(interpreter.getState() == USCXML_INSTANTIATED);
-            assert(interpreter.step() == USCXML_INITIALIZED);
+			assert(interpreter.getState() == USCXML_INSTANTIATED);
+			assert(interpreter.step() == USCXML_INITIALIZED);
 			assert(interpreter.step() == USCXML_IDLE);
 			assert(interpreter.step(true) == USCXML_MACROSTEPPED);
-            assert(interpreter.step() == USCXML_MICROSTEPPED);
+			assert(interpreter.step() == USCXML_MICROSTEPPED);
 			assert(interpreter.step() == USCXML_FINISHED);
 		}
 	}
